+++ src/main/java/com/deviget/minesweeper/infra/service/CellOperation.java
@@ -3,5 +3,6 @@ package com.deviget.minesweeper.infra.service;
 public enum CellOperation {
   UNCOVERING,
   FLAG_PLACEMENT,
-  QUESTION_MARK_PLACEMENT
+  QUESTION_MARK_PLACEMENT,
+  CLEARANCE
 }